devices/ophir: Fix serial lookup; expose range setting and serial numbers
Either the return type of the serial number query changed
between driver versions or the code never worked in the
first place. Either way, this fixes it and adds a driver
version check to make sure the COM object is at least as
recent as the one I tested with.

Also adds options to modify the power range.

This also removes the channel argument from the
wavelength-related functions, as all the functions exposed
by the class conceptually operate on a channel. This doesn't
make a big difference as all our power meters are only
single-channel anyway, but at least this gets rid of a few
magic constants (0).