Follow artiq.protocols -> sipyco split
In upstream ARTIQ, the artiq.protocols package was separated out
into a standalone project named "SiPyCo". This is a welcome development,
as it avoids dragging in all sorts of unnecessary dependencies on e.g.
a BeagleBone or Raspberry Pi used for interfacing with hardware, but
requires fairly widespread changes to imports.

The changes in this commit were made mechanically and have not yet
been tested in all instances.