devices: Add hacky stabilize_current_sense Ethernet interfacing code
Mostly Dougal's code.

This works (for setting feedback/feedforward coefficients, etc.), but
is in a somewhat messy state. (I would have put this onto a branch to
be revisited/squash-merged, but accidentally committed the setuptools
change in the previous change not realising that the code itself
wasn't part of the tree yet.)