SURF: add DynamicSplit support
This solver penalises high voltage slew rates and uses a different
low-jerk trajectory.

The start separation should be the measured value in the potential at
the start of the splitting waveform.

To make the critical point coincide with the zero-jerk crossing point,
the start curvature and end separation must be adjusted appropriately.