devices/tti_ql355: Retry identify() to work around device firmware bug
As per the comment, one our our supplies in lab one that otherwise
appears to work fine sometimes responds with empty lines.

We are not quite sure what the available workarounds for this would be;
e.g. is the buggy behaviour in the firmware somehow related to the
timing of commands sent? We have been running with this change for a few
months, and it appears to at least make the communication somewhat more
stable.